+++ lib/extension_api.dart
@@ -1,8 +1,7 @@
 library flutter_map_marker_popup.extension_api;
 
-// This was added for flutter_map_marker_cluster to add popup functionality.
-
-export 'src/popup_snap.dart';
-export 'src/popup_controller.dart';
-export 'src/marker_popup.dart';
+export 'src/animated_popup_container.dart';
 export 'src/popup_builder.dart';
+export 'src/popup_controller.dart';
+export 'src/popup_snap.dart';
+export 'src/simple_popup_container.dart';