+++ lib/flutter_map_marker_popup.dart
@@ -1,5 +1,6 @@
 library flutter_map_marker_popup;
 
+export 'src/popup_animation.dart';
 export 'src/popup_controller.dart';
 export 'src/popup_marker_layer.dart';
 export 'src/popup_marker_layer_options.dart';